app/requirements.txt
argon2-cffi
asyncpg
email-validator
fastapi
fastapi-mail
multipart
pillow
psycopg2
pydantic
pydantic-settings
PyJWT
rembg
SQLAlchemy
uvicorn
websockets
